✨ feat: Add Instagram Graph API module with postComment method
This commit adds a new module for the Instagram Graph API that includes a postComment method. It also includes a types.ts file defining the Config interface, which is used to configure the InstagramGraphAPI class. The postComment method makes a POST request to the comments endpoint of the Instagram Graph API, passing in the media ID and text as parameters. If successful, it returns a JSON response from the API.